# api-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/profiles/import/preview": {
      "post": {
        "description": "Upload a CSV file with a mapping config to see what profiles/memories would be created. Returns first 100 rows.",
        "operationId": "ImportV2Controller_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV file"
                  },
                  "mapping": {
                    "type": "string",
                    "description": "MappingConfig as JSON string"
                  }
                },
                "required": [
                  "file",
                  "mapping"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preview result with profiles, memories, errors and stats"
          }
        },
        "summary": "Preview a CSV import without writing to DB",
        "tags": [
          "Profiles — Bulk Import v2"
        ]
      }
    },
    "/v1/profiles/import": {
      "post": {
        "description": "Upload a CSV and mapping config. Returns a jobId for async processing.",
        "operationId": "ImportV2Controller_startImport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV file"
                  },
                  "mapping": {
                    "type": "string",
                    "description": "MappingConfig as JSON string"
                  }
                },
                "required": [
                  "file",
                  "mapping"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Import job queued",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "jobId": "uuid-here",
                    "status": "PROCESSING"
                  }
                }
              }
            }
          }
        },
        "summary": "Start a bulk profile import",
        "tags": [
          "Profiles — Bulk Import v2"
        ]
      }
    },
    "/v1/profiles/import/{jobId}": {
      "get": {
        "operationId": "ImportV2Controller_getJobStatus",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "description": "Import job UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job status with progress and stats",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "status": "PROCESSING",
                    "progress": 0.75,
                    "stats": {
                      "profileCount": 75,
                      "memoryCount": 60,
                      "errorCount": 2
                    },
                    "errors": []
                  }
                }
              }
            }
          },
          "404": {
            "description": "Job not found"
          }
        },
        "summary": "Get import job status",
        "tags": [
          "Profiles — Bulk Import v2"
        ]
      }
    }
  },
  "info": {
    "title": "Engram API",
    "description": "Memory infrastructure for AI agents",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-API-Key"
      },
      "user-id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AM-User-ID"
      }
    },
...
  }
}